Type Definition atsamd21g::adc::inputctrl::MUXNEG_R

type MUXNEG_R = R<u8, MUXNEG_A>;

Reader of field MUXNEG

Implementations

impl MUXNEG_R

pub fn variant(&self) -> Variant<u8, MUXNEG_A>

Get enumerated values variant

pub fn is_pin0(&self) -> bool

Checks if the value of the field is PIN0

pub fn is_pin1(&self) -> bool

Checks if the value of the field is PIN1

pub fn is_pin2(&self) -> bool

Checks if the value of the field is PIN2

pub fn is_pin3(&self) -> bool

Checks if the value of the field is PIN3

pub fn is_pin4(&self) -> bool

Checks if the value of the field is PIN4

pub fn is_pin5(&self) -> bool

Checks if the value of the field is PIN5

pub fn is_pin6(&self) -> bool

Checks if the value of the field is PIN6

pub fn is_pin7(&self) -> bool

Checks if the value of the field is PIN7

pub fn is_gnd(&self) -> bool

Checks if the value of the field is GND

pub fn is_iognd(&self) -> bool

Checks if the value of the field is IOGND