Type Definition atsamd21g::dmac::chctrlb::EVACT_R

type EVACT_R = R<u8, EVACT_A>;

Reader of field EVACT

Implementations

impl EVACT_R

pub fn variant(&self) -> Variant<u8, EVACT_A>

Get enumerated values variant

pub fn is_noact(&self) -> bool

Checks if the value of the field is NOACT

pub fn is_trig(&self) -> bool

Checks if the value of the field is TRIG

pub fn is_ctrig(&self) -> bool

Checks if the value of the field is CTRIG

pub fn is_cblock(&self) -> bool

Checks if the value of the field is CBLOCK

pub fn is_suspend(&self) -> bool

Checks if the value of the field is SUSPEND

pub fn is_resume(&self) -> bool

Checks if the value of the field is RESUME

pub fn is_sskip(&self) -> bool

Checks if the value of the field is SSKIP