atsamd21g::nvmctrl::ctrla

Type Alias CmdR

pub type CmdR = FieldReader<Cmdselect>;

Field CMD reader - Command

Aliased Type

struct CmdR { /* private fields */ }

Implementations

impl CmdR

pub const fn variant(&self) -> Option<Cmdselect>

Get enumerated values variant

pub fn is_er(&self) -> bool

Erase Row - Erases the row addressed by the ADDR register.

pub fn is_wp(&self) -> bool

Write Page - Writes the contents of the page buffer to the page addressed by the ADDR register.

pub fn is_ear(&self) -> bool

Erase Auxiliary Row - Erases the auxiliary row addressed by the ADDR register. This command can be given only when the security bit is not set and only to the user configuration row.

pub fn is_wap(&self) -> bool

Write Auxiliary Page - Writes the contents of the page buffer to the page addressed by the ADDR register. This command can be given only when the security bit is not set and only to the user configuration row.

pub fn is_sf(&self) -> bool

Security Flow Command

pub fn is_wl(&self) -> bool

Write lockbits

pub fn is_lr(&self) -> bool

Lock Region - Locks the region containing the address location in the ADDR register.

pub fn is_ur(&self) -> bool

Unlock Region - Unlocks the region containing the address location in the ADDR register.

pub fn is_sprm(&self) -> bool

Sets the power reduction mode.

pub fn is_cprm(&self) -> bool

Clears the power reduction mode.

pub fn is_pbc(&self) -> bool

Page Buffer Clear - Clears the page buffer.

pub fn is_ssb(&self) -> bool

Set Security Bit - Sets the security bit by writing 0x00 to the first byte in the lockbit row.

pub fn is_invall(&self) -> bool

Invalidates all cache lines.