atsamd21g::nvmctrl::ctrlb

Type Alias ReadmodeW

pub type ReadmodeW<'a, REG> = FieldWriter<'a, REG, 2, Readmodeselect>;

Field READMODE writer - NVMCTRL Read Mode

Aliased Type

struct ReadmodeW<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> ReadmodeW<'a, REG>
where REG: Writable + RegisterSpec, REG::Ux: From<u8>,

pub fn no_miss_penalty(self) -> &'a mut W<REG>

The NVM Controller (cache system) does not insert wait states on a cache miss. Gives the best system performance.

pub fn low_power(self) -> &'a mut W<REG>

Reduces power consumption of the cache system, but inserts a wait state each time there is a cache miss. This mode may not be relevant if CPU performance is required, as the application will be stalled and may lead to increase run time.

pub fn deterministic(self) -> &'a mut W<REG>

The cache system ensures that a cache hit or miss takes the same amount of time, determined by the number of programmed flash wait states. This mode can be used for real-time applications that require deterministic execution timings.