Type Definition atsamd21g::sercom0::usart::ctrla::MODE_R

type MODE_R = R<u8, MODE_A>;

Reader of field MODE

Implementations

impl MODE_R

pub fn variant(&self) -> Variant<u8, MODE_A>

Get enumerated values variant

pub fn is_usart_ext_clk(&self) -> bool

Checks if the value of the field is USART_EXT_CLK

pub fn is_usart_int_clk(&self) -> bool

Checks if the value of the field is USART_INT_CLK

pub fn is_spi_slave(&self) -> bool

Checks if the value of the field is SPI_SLAVE

pub fn is_spi_master(&self) -> bool

Checks if the value of the field is SPI_MASTER

pub fn is_i2c_slave(&self) -> bool

Checks if the value of the field is I2C_SLAVE

pub fn is_i2c_master(&self) -> bool

Checks if the value of the field is I2C_MASTER