Type Definition atsamd21g::sysctrl::bod33::PSEL_R

type PSEL_R = R<u8, PSEL_A>;

Reader of field PSEL

Implementations

impl PSEL_R

pub fn variant(&self) -> PSEL_A

Get enumerated values variant

pub fn is_div2(&self) -> bool

Checks if the value of the field is DIV2

pub fn is_div4(&self) -> bool

Checks if the value of the field is DIV4

pub fn is_div8(&self) -> bool

Checks if the value of the field is DIV8

pub fn is_div16(&self) -> bool

Checks if the value of the field is DIV16

pub fn is_div32(&self) -> bool

Checks if the value of the field is DIV32

pub fn is_div64(&self) -> bool

Checks if the value of the field is DIV64

pub fn is_div128(&self) -> bool

Checks if the value of the field is DIV128

pub fn is_div256(&self) -> bool

Checks if the value of the field is DIV256

pub fn is_div512(&self) -> bool

Checks if the value of the field is DIV512

pub fn is_div1k(&self) -> bool

Checks if the value of the field is DIV1K

pub fn is_div2k(&self) -> bool

Checks if the value of the field is DIV2K

pub fn is_div4k(&self) -> bool

Checks if the value of the field is DIV4K

pub fn is_div8k(&self) -> bool

Checks if the value of the field is DIV8K

pub fn is_div16k(&self) -> bool

Checks if the value of the field is DIV16K

pub fn is_div32k(&self) -> bool

Checks if the value of the field is DIV32K

pub fn is_div64k(&self) -> bool

Checks if the value of the field is DIV64K